+++ configs/retinanet_r50_fpn_1x.py
@@ -3,7 +3,7 @@ model = dict(
     type='RetinaNet',
     pretrained='modelzoo://resnet50',
     backbone=dict(
-        type='resnet',
+        type='ResNet',
         depth=50,
         num_stages=4,
         out_indices=(0, 1, 2, 3),
